+++ src/main/java/com/furnitureapp/repository/sales/CartRepository.java
@@ -12,7 +12,7 @@ import com.furnitureapp.repository.IRepository;
 
 import java.util.Set;
 
-public interface CartRepository extends IRepository<Cart, Long> {
+public interface CartRepository extends IRepository<Cart, Integer> {
 
     //Creation of the set that will contain the Cart objects and on which the CRUD method will be applied
         Set<Cart>list();
